Create Largest, Smallest number in an array of numbers (Python)
Simple logic to determine the smallest and the largest number from an array of numbers. The list is taken as an input using the split() function, hence the elements get stored as strings. Later during comparision, the individual elements are converted into int insitu and the smallest and largest numbers are found.